Initial layout/styling
Added a few boilerplate pages for category pages
Added search bar to navbar, as well as search bar component. Still need
to find a way to hook it up to data to comb through, as well as create
a search results page.
Minor changes to font sizes and backgrounds, including a light back-
ground for the body instead of dark